Refactor PR #28: fix turbo tasks, rename app.providers, add missing tests
- turbo.json: add lint task; fix test/typecheck dependsOn (test now
  depends on ^build, typecheck has no deps to avoid incorrect serialization)
- apps/mobile: rename app.providers/ → providers/ to remove unconventional
  dot in directory name; update _layout.tsx import and tsconfig include
- apps/desktop: export ShellCard through shared/index.ts barrel so app/
  layer imports from the layer boundary rather than reaching into internals
- packages/{sync,db,editor}: add "test" script so pnpm -r test covers them
- packages/{sync,db,editor}: add index.test.ts with initial coverage
- packages/core: strengthen definePlugin test — assert provides field and
  referential equality so identity contract is verified
