+++ apps/mobile/app/_layout.tsx
@@ -2,7 +2,7 @@ import "../global.css";
 
 import { Stack } from "expo-router";
 
-import { AppProviders } from "../app.providers/AppProviders";
+import { AppProviders } from "../providers/AppProviders";
 
 export function RootLayout() {
   return (
