+++ packages/db/package.json
@@ -10,6 +10,7 @@
     "@grove/core": "workspace:*"
   },
   "scripts": {
+    "test": "vitest run",
     "typecheck": "tsc -p tsconfig.json"
   }
 }
