+++ packages/db/src/index.test.ts
@@ -1,0 +1,9 @@
+import { describe, expect, it } from "vitest";
+
+import { createDatabasePlaceholder } from "./index";
+
+describe("createDatabasePlaceholder", () => {
+  it("returns idle as the initial database health", () => {
+    expect(createDatabasePlaceholder()).toBe("idle");
+  });
+});
