+++ packages/editor/src/index.test.ts
@@ -1,0 +1,9 @@
+import { describe, expect, it } from "vitest";
+
+import { createEditorMode } from "./index";
+
+describe("createEditorMode", () => {
+  it("returns markdown as the initial editor mode", () => {
+    expect(createEditorMode()).toBe("markdown");
+  });
+});
